chore: move back EditingRedPacketViewModel out of file
